refactor on checkTx - work in progress
