+++ packages/xchain-thorchain-query/src/thorchain-query.ts
@@ -1,6 +1,5 @@
 // import { Network } from '@xchainjs/xchain-client'
 
-import { LastBlock, ObservedTx, ObservedTxStatusEnum, TxOutItem } from '@xchainjs/xchain-thornode'
 import {
   Asset,
   AssetAtom,
@@ -8,9 +7,7 @@ import {
   AssetBTC,
   AssetRuneNative,
   Chain,
-  THORChain,
   assetAmount,
-  assetFromString,
   assetToBase,
   assetToString,
   baseAmount,
@@ -39,12 +36,10 @@ import {
   SwapEstimate,
   TotalFees,
   TxDetails,
-  TxStage,
-  TxStatus,
   UnitData,
 } from './types'
 import { getLiquidityProtectionData, getLiquidityUnits, getPoolShare, getSlipOnLiquidity } from './utils/liquidity'
-import { calcNetworkFee, getChain, getChainAsset } from './utils/swap'
+import { calcNetworkFee, getChainAsset } from './utils/swap'
 
 const BN_1 = new BigNumber(1)
 
@@ -482,213 +477,6 @@ export class ThorchainQuery {
     let minBlocks = runeValue.assetAmount.amount().toNumber() / txOutDelayRate
     minBlocks = minBlocks > maxTxOutOffset ? maxTxOutOffset : minBlocks
     return minBlocks * thorChainblocktime
-  }
-
-  /**
-   * For a given in Tx Hash (as returned by THORChainAMM.DoSwap()), finds the status of any THORChain transaction
-   * This function should be polled.
-   * @param
-   * @param inboundTxHash - needed to determine transactions stage
-   * @param sourceChain - extra parameter
-   * @returns - object tx status
-   */
-  public async checkTx(inboundTxHash: string, sourceChain?: Chain): Promise<TxStatus> {
-    let txStatus: TxStatus = { stage: TxStage.INBOUND_CHAIN_UNCONFIRMED, seconds: 0 }
-    const txData = await this.thorchainCache.thornode.getTxData(inboundTxHash)
-    const scheduledQueueItem = (await this.thorchainCache.thornode.getscheduledQueue()).find(
-      (item: TxOutItem) => item.in_hash === inboundTxHash,
-    )
-    //console.log(`Tx stage ${txStatus.stage}\nTx seconds left ${txStatus.seconds}`)
-    // Check to see if the transaction has been observed
-    if (txData.observed_tx == undefined) {
-      txStatus = await this.checkTxDefined(txStatus, sourceChain)
-      return txStatus
-    }
-    // If its scheduled and observed
-    if (scheduledQueueItem && txData.observed_tx) {
-      txStatus = await this.checkObservedOnly(txStatus, scheduledQueueItem, txData.observed_tx, sourceChain)
-    }
-    //console.log(`Tx stage ${txStatus.stage}\nTx seconds left ${txStatus.seconds}`)
-
-    // Retrieve asset and chain from memo
-    const pool = txData.observed_tx.tx.memo?.split(`:`)
-    if (!pool) throw Error(`No pool found from memo`)
-    const getAsset = assetFromString(pool[1].toUpperCase())
-    if (!getAsset) throw Error(`Invalid pool asset`)
-
-    // Retrieve thorchain blockHeight for the tx
-    if (!txData.observed_tx.tx?.id) throw Error('No action observed')
-    const recordedAction = await this.thorchainCache.midgard.getActions('', txData.observed_tx.tx.id)
-    const recordedTCBlock = recordedAction.find((block) => {
-      return block
-    })
-    if (!recordedTCBlock?.height) throw Error('No recorded block height')
-
-    // Retrieve thorchains last observed block height
-    const lastBlock = await this.thorchainCache.thornode.getLastBlock()
-    const lastBlockHeight = lastBlock.find((obj) => obj.chain === getAsset?.chain)
-
-    // Check to see if its in the outbound queue
-    if (scheduledQueueItem) {
-      txStatus = await this.checkOutboundQueue(txStatus, scheduledQueueItem, lastBlockHeight)
-      // Check to see if there is an outbound wait
-      if (scheduledQueueItem?.height != undefined && txStatus.stage < TxStage.OUTBOUND_CHAIN_CONFIRMED) {
-        txStatus = await this.checkOutboundTx(txStatus, scheduledQueueItem, lastBlockHeight)
-      }
-      //console.log(`Tx stage ${txStatus.stage}\nTx seconds left ${txStatus.seconds}`)
-      return txStatus
-    }
-
-    // If not in queue, outbound Tx sent // check synth // check it status == done
-    if (!scheduledQueueItem && getAsset) {
-      txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
-      if (getAsset?.synth) {
-        if (txData.observed_tx?.status == ObservedTxStatusEnum.Done) {
-          txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
-          txStatus.seconds = 0
-        } else {
-          txStatus.seconds = 6
-        }
-        //console.log(`Tx stage ${txStatus.stage}\nTx seconds left ${txStatus.seconds}`)
-        return txStatus
-      }
-      if (txData.observed_tx?.status == ObservedTxStatusEnum.Done && getAsset.chain != Chain.THORChain) {
-        // Retrieve recorded asset block height for the Outbound asset
-        const recordedBlockHeight = await this.thorchainCache.thornode.getLastBlock(+recordedTCBlock.height)
-        // Match outbound asset to block record
-        const assetBlockHeight = recordedBlockHeight.find((obj) => obj.chain === getAsset?.chain)
-        if (lastBlockHeight?.last_observed_in && assetBlockHeight?.last_observed_in) {
-          const chainblockTime = this.chainAttributes[getAsset.chain].avgBlockTimeInSecs
-          // Difference between current block and the recorded tx block for the outbound asset
-          const blockDifference = lastBlockHeight.last_observed_in - assetBlockHeight.last_observed_in
-          const timeElapsed = blockDifference * chainblockTime
-          // If the time elapsed since the tx is greater than the chains block time, assume tx has 1 ocnfirmation else return time left to wait
-          txStatus.seconds = timeElapsed > chainblockTime ? 0 : chainblockTime - timeElapsed
-          console.log(timeElapsed)
-        } else if (txData.observed_tx.tx.id && lastBlockHeight?.thorchain) {
-          const recordedAction = await this.thorchainCache.midgard.getActions(txData.observed_tx.tx.id)
-          const recordedBlockheight = recordedAction.find((block) => {
-            return block
-          })
-          if (!recordedBlockheight) throw Error(`No height recorded`)
-          const chainblockTime = this.chainAttributes[getAsset.chain].avgBlockTimeInSecs
-          console.log(chainblockTime)
-          const blockDifference = lastBlockHeight?.thorchain - +recordedBlockheight.height
-          console.log(blockDifference)
-          const timeElapsed =
-            (blockDifference * chainblockTime) / this.chainAttributes[getAsset.chain].avgBlockTimeInSecs
-          txStatus.seconds = timeElapsed > chainblockTime ? 0 : chainblockTime - timeElapsed
-          console.log(txStatus.seconds)
-          txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
-        }
-        //console.log(`Tx stage ${txStatus.stage}\nTx seconds left ${txStatus.seconds}`)
-        return txStatus
-      } else {
-        txStatus.seconds = 0
-        txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
-      }
-      //console.log(`Tx stage ${txStatus.stage}\nTx seconds left ${txStatus.seconds}`)
-      return txStatus
-    } else {
-      // case example "memo": "OUT:08BC062B248F6F27D0FECEF1650843585A1496BFFEAF7CB17A1CBC30D8D58F9C" where no asset is found its a thorchain tx. Confirms in ~6 seconds
-      txStatus.seconds = 0
-      txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
-    }
-    return txStatus
-  }
-
-  /** Stage 1  */
-  private async checkTxDefined(txStatus: TxStatus, sourceChain?: Chain): Promise<TxStatus> {
-    // If there is an error Thornode does not know about it. wait 60 seconds
-    // If a long block time like BTC, can check or poll to see if the status changes.
-    if (sourceChain) {
-      txStatus.seconds = this.chainAttributes[sourceChain].avgBlockTimeInSecs
-    } else {
-      txStatus.seconds = 60
-    }
-    return txStatus
-  }
-
-  /** Stage 2, THORNode has seen it. See if observed only (conf counting) or it has been processed by THORChain  */
-  // e.g. https://thornode.ninerealms.com/thorchain/tx/365AC447BE6CE4A55D14143975EE3823A93A0D8DE2B70AECDD63B6A905C3D72B
-  private async checkObservedOnly(
-    txStatus: TxStatus,
-    scheduledQueueItem: TxOutItem,
-    observed_tx?: ObservedTx,
-    sourceChain?: Chain,
-  ): Promise<TxStatus> {
-    if (observed_tx?.tx?.chain != undefined) {
-      sourceChain = getChain(observed_tx.tx.chain)
-    } else {
-      throw new Error(`Cannot get source chain ${observed_tx?.tx?.chain}`)
-    }
-
-    //If observed by not final, need to wait till the finalised block before moving to the next stage, blocks in source chain
-    if (observed_tx?.block_height && observed_tx?.finalise_height && scheduledQueueItem.height) {
-      if (observed_tx.block_height < observed_tx.finalise_height) {
-        txStatus.stage = TxStage.CONF_COUNTING
-        const blocksToWait = observed_tx.finalise_height - scheduledQueueItem.height
-        txStatus.seconds = blocksToWait * this.chainAttributes[sourceChain].avgBlockTimeInSecs
-      } else if (observed_tx.status != ObservedTxStatusEnum.Done) {
-        // processed but not yet full final, e.g. not 2/3 nodes signed
-        txStatus.seconds = this.chainAttributes[THORChain].avgBlockTimeInSecs // wait one more TC block
-        txStatus.stage = TxStage.TC_PROCESSING
-      }
-    }
-    return txStatus
-  }
-  /**
-   * Stage 3
-   * @param txStatus
-   * @param txData
-   * @param scheduledQueue
-   * @param scheduledQueueItem
-   * @param lastBlockHeight
-   * @returns
-   */
-  private async checkOutboundQueue(
-    txStatus: TxStatus,
-    scheduledQueueItem?: TxOutItem,
-    lastBlockHeight?: LastBlock,
-  ): Promise<TxStatus> {
-    // If the scheduled block is greater than the current block, need to wait that amount of blocks till outbound is sent
-    if (scheduledQueueItem?.height && lastBlockHeight?.thorchain) {
-      if (lastBlockHeight.thorchain < scheduledQueueItem?.height) {
-        const blocksToWait = scheduledQueueItem.height - lastBlockHeight.thorchain
-        txStatus.stage = TxStage.OUTBOUND_QUEUED
-        txStatus.seconds = blocksToWait * this.chainAttributes[THORChain].avgBlockTimeInSecs
-        return txStatus
-      } else {
-        txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
-        return txStatus
-      }
-    }
-    return txStatus
-  }
-  /** Stage 4 */
-  private async checkOutboundTx(
-    txStatus: TxStatus,
-    scheduledQueueItem?: TxOutItem,
-    lastBlockHeight?: LastBlock,
-  ): Promise<TxStatus> {
-    if (scheduledQueueItem?.height && lastBlockHeight?.thorchain) {
-      const blockDifference = scheduledQueueItem.height - lastBlockHeight?.thorchain
-      const timeElapsed = blockDifference * this.chainAttributes[THORChain].avgBlockTimeInSecs
-      if (blockDifference == 0) {
-        // If Tx has just been sent, Stage 3 should pick this up really
-        txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
-        txStatus.seconds = this.chainAttributes[THORChain].avgBlockTimeInSecs
-      } else if (timeElapsed < txStatus.seconds) {
-        // if the time passed since the outbound TX was sent is less than the outbound block time, outbound Tx unconfirmed, wait a bit longer.
-        txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
-        txStatus.seconds = this.chainAttributes[THORChain].avgBlockTimeInSecs - timeElapsed // workout how long to wait
-      } else {
-        // time passed is greater than outbound Tx time, Tx is confirmed. Thus stage 5
-        txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
-        txStatus.seconds = 0
-      }
-    }
-    return txStatus
   }
 
   /**
